+++ sdk/python/lib/pulumi/runtime/stack.py
@@ -21,7 +21,7 @@ from inspect import isawaitable
 from typing import Callable, Any, Dict, List
 
 from ..resource import ComponentResource, Resource, ResourceTransformation
-from .settings import get_project, get_stack, get_root_resource, set_root_resource
+from .settings import get_project, get_stack, get_root_resource, is_dry_run, set_root_resource
 from .rpc_manager import RPC_MANAGER
 from .. import log
 
@@ -148,6 +148,17 @@ def massage(attr: Any, seen: List[Any]):
 
     if isawaitable(attr):
         return Output.from_input(attr).apply(lambda v: massage(v, seen))
+
+    if isinstance(attr, Resource):
+        result = massage(attr.__dict__, seen)
+
+        # In preview only, we mark the result with "@isPulumiResource" to indicate that it is derived
+        # from a resource. This allows the engine to perform resource-specific filtering of unknowns
+        # from output diffs during a preview. This filtering is not necessary during an update because
+        # all property values are known.
+        if is_dry_run():
+            result["@isPulumiResource"] = True
+        return result
 
     if hasattr(attr, "__dict__"):
         # recurse on the dictionary itself.  It will be handled above.